Move the scan function logic back into ble.service
